+++ IsekaiMod/Content/Features/IsekaiSuccubus/SuccubusWingsAbility.cs
@@ -15,18 +15,20 @@ using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Formations.Facts;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components.AreaEffects;
+using UnityEngine;
 
 namespace IsekaiMod.Content.Features.IsekaiSuccubus
 {
     class SuccubusWingsAbility
     {
+        // Exclude Mutagen wings
+        private static readonly BlueprintBuff BuffWingsMutagen = Resources.GetBlueprint<BlueprintBuff>("e4979934bdb39d842b28bee614606823");
+
+        // Icon
+        private static readonly Sprite Icon_Wings = Resources.GetBlueprint<BlueprintBuff>("4113178a8d5bf4841b8f15b1b39e004f").m_Icon;
+
         public static void Add()
         {
-            // Buff
-            var BuffWingsMutagen = Resources.GetBlueprint<BlueprintBuff>("e4979934bdb39d842b28bee614606823");
-
-            // Ability
-            var Icon_Wings = Resources.GetBlueprint<BlueprintBuff>("4113178a8d5bf4841b8f15b1b39e004f").m_Icon;
             var SuccubusWingsBuff = Helpers.CreateBlueprint<BlueprintBuff>("SuccubusWingsBuff", bp => {
                 bp.SetName("Succubus Wings");
                 bp.SetDescription("You gain a pair of wings that grant a +3 dodge {g|Encyclopedia:Bonus}bonus{/g} to {g|Encyclopedia:Armor_Class}AC{/g} "
@@ -90,27 +92,25 @@ namespace IsekaiMod.Content.Features.IsekaiSuccubus
             });
 
             // Patch pit spells
-            var PitOfDespairArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("b905a3c987f22cb49a246f0ab211f34c");
-            var TricksterRecreationalPitArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("bf68ec704dc186549a7c6fbf22d3d661");
-            var CreatePitArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("cf742a1d377378e4c8799f6a3afff1ba");
-            var SpikedPitArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("beccc33f543b1f8469c018982c23ac06");
-            var AcidPitArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("e122151e93e44e0488521aed9e51b617");
-            var HungryPitArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("d086b1aeb367a5b43808d34c321955d1");
-            var RiftOfRuinArea = Resources.GetBlueprint<BlueprintAbilityAreaEffect>("9b51157a5305dbf4184bf15bdad39226");
-            PitOfDespairArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = PitOfDespairArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            TricksterRecreationalPitArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = TricksterRecreationalPitArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            CreatePitArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = CreatePitArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            SpikedPitArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = SpikedPitArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            AcidPitArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = AcidPitArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            HungryPitArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = HungryPitArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
-            RiftOfRuinArea.GetComponent<AreaEffectPit>().m_EffectsImmunityFacts = RiftOfRuinArea.GetComponent<AreaEffectPit>()
-                .m_EffectsImmunityFacts.AddToArray(SuccubusWingsBuff.ToReference<BlueprintUnitFactReference>());
+            PatchPitAreaEffects(SuccubusWingsBuff);
+        }
+        private static void PatchPitAreaEffects(BlueprintBuff wingsBuff)
+        {
+            BlueprintAbilityAreaEffect[] pitAbilityAreaEffects = new BlueprintAbilityAreaEffect[]
+            {
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("b905a3c987f22cb49a246f0ab211f34c"), // PitOfDespairArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("bf68ec704dc186549a7c6fbf22d3d661"), // TricksterRecreationalPitArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("cf742a1d377378e4c8799f6a3afff1ba"), // CreatePitArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("beccc33f543b1f8469c018982c23ac06"), // SpikedPitArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("e122151e93e44e0488521aed9e51b617"), // AcidPitArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("d086b1aeb367a5b43808d34c321955d1"), // HungryPitArea
+                Resources.GetBlueprint<BlueprintAbilityAreaEffect>("9b51157a5305dbf4184bf15bdad39226")  // RiftOfRuinArea
+            };
+            foreach (BlueprintAbilityAreaEffect pitAbilityAreaEffect in pitAbilityAreaEffects)
+            {
+                var areaEffectPit = pitAbilityAreaEffect.GetComponent<AreaEffectPit>();
+                areaEffectPit.m_EffectsImmunityFacts = areaEffectPit.m_EffectsImmunityFacts.AddToArray(wingsBuff.ToReference<BlueprintUnitFactReference>());
+            }
         }
     }
 }